requirements.txt
claude-agent-sdk
aiogram~=3.26
python-dotenv
anyio
